replace WhatIs to components
