+++ src/pages/index.js
@@ -11,6 +11,7 @@ import styles from './styles.module.css';
 import { Grid } from '@mui/material';
 import SupportersList from '../components/supportersList';
 import Button from '../components/button'
+import WhatIs from '../components/whatIs'
 
 export default function Home() {
   const context = useDocusaurusContext();
@@ -85,27 +86,3 @@ export default function Home() {
     </Layout>
   );
 }
-
-const WhatIs = () => (
-  <div className={clsx('hero', styles.hero)}>
-    <div className="container">
-      <div className="row">
-        <div className="col">
-          <h1 className="text-center title"><Translate>What is Karmada?</Translate></h1>
-          <p className="hero__subtitle">
-            <small>
-              <Translate>
-              Karmada (Kubernetes Armada) is a Kubernetes management system that enables you to run your cloud-native applications across multiple Kubernetes clusters and clouds, with no changes to your applications. By speaking Kubernetes-native APIs and providing advanced scheduling capabilities, Karmada enables truly open, multi-cloud Kubernetes.
-              </Translate>
-              <br />
-              <br />
-              <Translate>
-              Karmada aims to provide turnkey automation for multi-cluster application management in multi-cloud and hybrid cloud scenarios, with key features such as centralized multi-cloud management, high availability, failure recovery, and traffic scheduling.
-              </Translate>
-            </small>
-          </p>
-        </div>
-      </div>
-    </div>
-  </div>
-);
